requirements.txt
annotated-types==0.6.0
anyio==4.10.0
asttokens==3.0.1
Bottleneck==1.4.2
certifi==2026.1.4
click==8.2.1
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
executing==2.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
mkl_fft==2.1.1
mkl_random==1.3.0
mkl-service==2.5.2
nest_asyncio==1.6.0
numexpr==2.14.1
numpy==2.3.5
openai==2.14.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pip==25.3
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==27.1.0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
threadpoolctl==3.5.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
wcwidth==0.2.14
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pyjwt==2.11.0
